# app/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IT system to operate a clinic network",
        "version": "0.6.2",
        "title": "Clinic network system",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "michal.balicki@onet.pl"
        }
    },
    "host": "https://komponentowe-backend.herokuapp.com/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/reception/:login": {
            "get": {
                "tags": [
                    "reception"
                ],
                "summary": "Reception worker found by login",
                "description": "Requires apikey",
                "operationId": "reception",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginString"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/reception/signup": {
            "post": {
                "tags": [
                    "reception"
                ],
                "summary": "Reception worker register",
                "description": "Requires apikey",
                "operationId": "receptionSignup",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/doctorSignup"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/reception/datafillup": {
            "put": {
                "tags": [
                    "reception"
                ],
                "summary": "Reception worker profile data fill up",
                "description": "Requires apikey",
                "operationId": "reception",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/doctorDataFillUp"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/reception/data/:id": {
            "get": {
                "tags": [
                    "reception"
                ],
                "summary": "Reception worker profile data",
                "description": "Requires apikey",
                "operationId": "receptionData",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/profileData"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/reception/login": {
            "post": {
                "tags": [
                    "reception"
                ],
                "summary": "Reception worker login",
                "description": "Requires apikey",
                "operationId": "receptionLogin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/login"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/loginReceptionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "content": {}
                    }
                }
            }
        },
        "/reception/:id": {
            "delete": {
                "tags": [
                    "reception"
                ],
                "summary": "Delete reception worker by worker's id",
                "description": "Requires apikey",
                "operationId": "receptionDelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/login"
                            }
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "loginString": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                }
            }
        },
        "doctorSignup": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "haslo": {
                    "type": "string"
                }
            }
        },
        "profileData": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "haslo": {
                    "type": "string"
                },
                "imie": {
                    "type": "string"
                },
                "nazwisko": {
                    "type": "string"
                },
                "idplci": {
                    "type": "string"
                },
                "ulica": {
                    "type": "string"
                },
                "numerlokalu": {
                    "type": "string"
                },
                "miejscowosc": {
                    "type": "string"
                },
                "kodpocztowy": {
                    "type": "string"
                },
                "numertelefonu": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "login": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "haslo": {
                    "type": "string"
                }
            }
        },
...
        "loginReceptionResponse": {
            "type": "object",
            "properties": {
                "receptionistid": {
                    "type": "integer"
                },
                "workerid": {
                    "type": "integer"
                }
            }
        }
    }
}